refactor: move stats into one file
