+++ src/server/bitops_family.cc
@@ -16,6 +16,7 @@
 #include "server/command_families.h"
 #include "server/command_registry.h"
 #include "server/conn_context.h"
+#include "server/db_slice.h"
 #include "server/engine_shard_set.h"
 #include "server/error.h"
 #include "server/namespaces.h"
