+++ src/server/cluster/cluster_family.cc
@@ -23,7 +23,9 @@
 #include "server/cluster/coordinator.h"
 #include "server/command_registry.h"
 #include "server/conn_context.h"
+#include "server/db_slice.h"
 #include "server/dflycmd.h"
+#include "server/engine_shard_set.h"
 #include "server/error.h"
 #include "server/journal/journal.h"
 #include "server/main_service.h"
