+++ src/server/detail/save_stages_controller.cc
@@ -12,7 +12,9 @@
 #include "base/flags.h"
 #include "base/logging.h"
 #include "core/detail/gen_utils.h"
+#include "server/db_slice.h"
 #include "server/detail/snapshot_storage.h"
+#include "server/engine_shard_set.h"
 #include "server/main_service.h"
 #include "server/namespaces.h"
 #include "server/script_mgr.h"
