+++ src/server/replica.cc
@@ -27,6 +27,7 @@ extern "C" {
 #include "facade/redis_parser.h"
 #include "facade/reply_capture.h"
 #include "facade/socket_utils.h"
+#include "server/engine_shard_set.h"
 #include "server/error.h"
 #include "server/journal/executor.h"
 #include "server/journal/journal.h"
