+++ src/server/search/search_family.cc
@@ -32,6 +32,7 @@
 #include "server/config_registry.h"
 #include "server/conn_context.h"
 #include "server/container_utils.h"
+#include "server/db_slice.h"
 #include "server/engine_shard_set.h"
 #include "server/namespaces.h"
 #include "server/search/aggregator.h"
